Safari-611.1.10/LayoutTests/fast/dom/Element/body-scrollHeight-basics-quirks-expected.txt

Test the "scrollHeight" properties on the body element in quirks mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial values
PASS document.body.scrollHeight is document.documentElement.scrollHeight

Now we create a body element, it should not have a scrollHeight since it is not the first body element of the document.
PASS document.body.scrollHeight is document.documentElement.scrollHeight
PASS floatingBody.scrollHeight is 0

We add a new body element after the first body element, this should have no impact on the state.
PASS document.body.scrollHeight is document.documentElement.scrollHeight
PASS secondSiblingBody.scrollHeight is secondSiblingBody.clientHeight
PASS document.body.scrollHeight is not secondSiblingBody.scrollHeight

We add a new body element inside the first body element, this should have no impact on the state.
PASS document.body.scrollHeight is document.documentElement.scrollHeight
PASS bodyChildOfMainBody.scrollHeight is bodyChildOfMainBody.clientHeight
PASS document.body.scrollHeight is not bodyChildOfMainBody.scrollHeight

Getting the scrollHeight should not require a renderer.
PASS document.body.style.display = 'none' is 'none'
PASS document.body.scrollHeight is document.documentElement.scrollHeight
PASS successfullyParsed is true

TEST COMPLETE
